Changed input validation
checking file paths as initial input validation seemed pretty redundant because the script would throw a "file not found" exception anyway once it called "Get-ChildItems" on the specified path. This form of input validation was removed.

Also, it was noticed that since none of the file type switches are flagged as mandatory (because any one of them can be selected, so they can't be mandatory) it was possible for the user to specify none of them. If none were specified, an exception would be thrown regarding a null object, which can be a bit cryptic. Input validation was added that if none were specified, it would default to "internet".